Add routes.ts file and move server files to server folder
